Remove mandatory clip from Scroll.kt
I turn the clip of scroll optional because I figure out a scenario using dragging gesture that needs to drag the items out container. Calling `clipScrollableContainer` modifier method every time made impossible to do that. So, the solution is turn a option with default argument as `true` to not affect the commons case.